Fix post sorting, image changes
For some reason, I made the posts sort by their slug instead of the date they were posted. I always intended to make them sort by date, but I did it by slug for some reason.
Changed settings so that the auto image format isn't being processed by default. It was sending the original format of the image as well as png, negating benefits of webp.
Changed the resolution of the blurred image placeholder to be a bit higher res.
Changed the layout of images to full width so it will hopefully change the size of the image based on its displayed width.